u32 NodeId/Term migration, direct-submit LocalWriter, mmap pre-fault and pooled preallocation, and per-connection response routing
Raft type system (type_config.rs, codec.rs, record_format.rs):
- Migrate NodeId and Term from u64 to u32 across the entire Raft type config,
  halving wire sizes: LeaderId 16→8 bytes, LogId 24→16, Vote stays 16 (more padding)
- Introduce BisqueNodeId/BisqueTerm/BisqueLeaderId type aliases in type_config.rs
- Codec impls switch from generic `C: RaftTypeConfig` bounds to concrete
  `LeaderId<u32,u32>` types — simpler, monomorphized, no trait gymnastics
- Membership node_id encoding shrinks from 8→4 bytes per entry
- AtomicVote and AtomicLogId updated for u32 term/node_id with u64 casts
  for bit-packing arithmetic

Mmap storage (storage_mmap.rs):
- Default segment size 1 MiB→64 MiB; max record size 1 MiB→4 MiB to
  accommodate coalesced TCP entries from ForwardAcceptor
- Entry header layout compressed: [term:4][node_id_tag:4][index:8] = 16 bytes
  (packs node_id bits 8..31 + tag bits 0..7 into one u32)
- madvise_write() and madvise_read() helpers with MADV_POPULATE_WRITE/READ
  on Linux for pre-faulting mmap pages (eliminates per-page faults on write)
- Segment mmap opens use .populate().no_reserve_swap() for eager page-in
- New config: disable_fsync (benchmark mode, callbacks fire without sync),
  disable_crc (skip CRC compute on writes), prealloc_pool_size
- Default fsync_delay 1ms→100ms for better write coalescing
- No-fsync mode: callbacks dispatched via unbounded mpsc channel to dedicated
  tokio task, fully decoupling commit latency from append
- CRC validation skips zero-CRC entries (written with disable_crc)

Write pipeline (write_batcher.rs, raft_writer.rs, forward.rs):
- LocalBatcher drain loop eliminated — LocalWriter now builds
  TAG_FORWARDED_BATCH commands and submits directly to RaftWriter MPMC queue
- Each LocalWriter is an independent proposer with its own backlog charge,
  removing the single-drain-loop bottleneck
- RaftWriter gains max_entries_per_proposal cap (default 64) to prevent
  giant proposals that serialise mmap writes; smaller proposals enable
  Raft core pipelining across workers
- RaftWriterStats extended with proposal_errors, entry_errors,
  error_released_bytes counters
- RaftWriter::new/with_backlog now generic over SM for state machine flexibility
- OutboundBuf lock-free partition buffers removed from forward.rs (replaced
  by direct LocalWriter→RaftWriter path); batch header and response framing
  retained for TCP forwarding
- Default max_raft_entry_bytes raised 512 KiB→1 MiB

Server handler (mq-server/handler.rs, bisque/mq.rs):
- MqRouter trait: get_batcher() replaced by create_submitter() which takes
  a ResponseCallback — each connection gets its own submitter and response
  channel instead of sharing a single batcher
- ConsumerHandler gains per-connection LocalSubmitter, PendingOp tracking
  map, and crossfire SPSC response channel for async response dispatch
- BisqueMqRouter: primary_batcher/partition_batchers replaced by
  SubmitterFactory (stores LocalWriter + ClientRegistry for lazy per-
  connection submitter creation)
- Full integration wiring: MqState::start_mq_group now creates RaftWriter,
  LocalBatcher, and passes async_apply ClientRegistry to the router

Async apply (async_apply.rs):
- ClientPartition reg_rx switched from tokio::mpsc::unbounded to
  crossfire::mpsc::bounded(128) for bounded backpressure on registration
- TAG_BATCH apply path removed (superseded by TAG_FORWARDED_BATCH with
  node_id=0 for local batches)
- Local batch (node_id=0) apply: merged single-pass loop that applies
  sub-commands and delivers rich ResponseEntry frames via ClientRegistry
- batch_registry oneshot map removed — no longer needed

Protocol adapters (mq-kafka, mq-mqtt, mq-sqs):
- Handler signatures updated for MqRouter::create_submitter API change

+++ crates/mq/src/raft.rs
@@ -30,7 +30,7 @@ use crate::types::MqCommand;
 /// - Actor rebalancing
 pub struct MqRaftNode {
     raft: Raft<MqTypeConfig>,
-    node_id: u64,
+    node_id: u32,
     group_id: u64,
     config: MqConfig,
     metadata: Option<Arc<MqMetadata>>,
@@ -42,7 +42,7 @@ pub struct MqRaftNode {
 }
 
 impl MqRaftNode {
-    pub fn new(raft: Raft<MqTypeConfig>, node_id: u64, config: MqConfig) -> Self {
+    pub fn new(raft: Raft<MqTypeConfig>, node_id: u32, config: MqConfig) -> Self {
         Self {
             raft,
             node_id,
@@ -217,7 +217,7 @@ impl MqRaftNode {
     where
         F: Fn(
                 Raft<MqTypeConfig>,
-                u64,
+                u32,
             ) -> std::pin::Pin<Box<dyn std::future::Future<Output = ()> + Send>>
             + Send
             + Sync
